S3 finished
now it upload the buzzcard image and save the image to the device.
The only issue is after it takes the picture, it has to be saved on the device. There is a limit on how big the data can be to be transferred between activities. the image exceeds the limit. So i have to save it, then read it from the signup activity.
this can be solve by chopping the image in to multiple segments and send transfer them separately. But for saving the headache, we just save it.